+++ __tests__/01_basic_spec.js
@@ -1,6 +1,7 @@
 /* eslint-env jest */
 
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import renderer from 'react-test-renderer';
 import { composeWithState } from '../src/index';